Address an issue with audio file location metadata
Address an issue in which the "z" elevation provided
in audio metadata was exported as:

z = 0.0 - firing_point.origin.height_above_elevation

rather than:

z = GPS.height_above_elevation - firing_point.origin.height_above_elevation

as intended.

As with the x and y values, these values derive from
the real-time location reported by the on-board GPS at
the timestamp associated with the start of each file.

Thanks to Dan Hack for pointing out the discrepancy in
the previously reported data.